Added support for parsing Slash Outgoing WebHook commands and slash commands style of the format
  /command action value
  /command action p1=v1; p2=v2; ...

An example of an outgoing WebHook or slash command using the first format is '/assistant query what's the weather?' where '/assistant' is the trigger word or slash command, 'query' is the action and 'what's the weather?' is the value.

In the second form, parameter values containing semi-colons can either escape the semi-colon using a backslash character, i.e '\;', or quote the value using single quotes or double quotes. An example of an outgoing WebHook or slash command using this format is '/appointment add title=doctor visit; time=Feb 3 2016 2 PM; location=Children's Hospital' where '/appointment' is the trigger word or slash command, 'add' is the action and title, time, and location are parameters.

This makes it much simpler to add support for Slack slash commands and outgoing webhook that take parameters from the user.

Updated the SlackSample to illustrate use of this for both outgoing webhooks and slash commands.